+++ packages/theme/src/theme.ts
@@ -1,5 +1,6 @@
 import { buttonTheme, ButtonTheme } from "./buttonTheme";
 import { headingTheme, HeadingTheme } from "./headingTheme";
+import { headerTheme, HeaderTheme } from "./headerTheme";
 import { textTheme, TextTheme } from "./textTheme";
 import { SlideOverTheme, slideOverTheme } from "./slideOverTheme";
 
@@ -8,6 +9,7 @@ export interface Theme {
   heading: HeadingTheme;
   text: TextTheme;
   slideOver: SlideOverTheme;
+  header: HeaderTheme;
 }
 
 export const defaultTheme: Theme = {
@@ -15,4 +17,5 @@ export const defaultTheme: Theme = {
   heading: headingTheme,
   text: textTheme,
   slideOver: slideOverTheme,
+  header: headerTheme,
 };