+++ packages/atoms/button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tility-ui/button
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @tility-ui/theme@0.2.1
+
 ## 0.1.2
 
 ### Patch Changes